Comorbidity mapping.

Maps comorbidity conditions using algorithms from the Charlson and the Elixhauser comorbidity scores.

comorbidity(x, id, code, map, assign0, labelled = TRUE, tidy.codes = TRUE)

Arguments

x
A tidy data.frame (or a data.table; tibbles are supported too) with one column containing an individual ID and a column containing all diagnostic codes. Extra columns other than ID and codes are discarded. Column names must be syntactically valid names, otherwise they are forced to be so by calling the make.names() function.

id
Column of x containing the individual ID.

code
Column of x containing diagnostic codes. Codes must be in upper case with no punctuation in order to be properly recognised.

map
The mapping algorithm to be used (values are case-insensitive). Possible values are the Charlson score with either ICD-10 or ICD-9-CM codes (charlson_icd10_quan, charlson_icd9_quan) and the Elixhauser score, again using either ICD-10 or ICD-9-CM (elixhauser_icd10_quan, elixhauser_icd9_quan). These mapping are based on the paper by Quan et al. (2011). It is also possible to obtain a Swedish (charlson_icd10_se) or Australian (charlson_icd10_am) modification of the Charlson score using ICD-10 codes.

assign0
Apply a hierarchy of comorbidities: should a comorbidity be present in a patient with different degrees of severity, then the milder form will be assigned a value of 0. By doing this, a type of comorbidity is not counted more than once in each patient. The comorbidities that are affected by this argument are:

"Mild liver disease" (mld) and "Moderate/severe liver disease" (msld) for the Charlson score;

"Diabetes" (diab) and "Diabetes with complications" (diabwc) for the Charlson score;

"Cancer" (canc) and "Metastatic solid tumour" (metacanc) for the Charlson score;

"Hypertension, uncomplicated" (hypunc) and "Hypertension, complicated" (hypc) for the Elixhauser score;

"Diabetes, uncomplicated" (diabunc) and "Diabetes, complicated" (diabc) for the Elixhauser score;

"Solid tumour" (solidtum) and "Metastatic cancer" (metacanc) for the Elixhauser score.

labelled
Attach labels to each comorbidity, compatible with the RStudio viewer via the utils::View() function. Defaults to TRUE.

tidy.codes
Tidy diagnostic codes? If TRUE, all codes are converted to upper case and all non-alphanumeric characters are removed using the regular expression [^[:alnum:]]. Defaults to TRUE.

Value

A data frame with id, columns relative to each comorbidity domain, comorbidity score, weighted comorbidity score, and categorisations of such scores, with one row per individual.

For the Charlson score, the following variables are included in the dataset:

The id variable as defined by the user;

ami, for acute myocardial infarction;

chf, for congestive heart failure;

pvd, for peripheral vascular disease;

cevd, for cerebrovascular disease;

dementia, for dementia;

copd, chronic obstructive pulmonary disease;

rheumd, for rheumatoid disease;

pud, for peptic ulcer disease;

mld, for mild liver disease;

diab, for diabetes without complications;

diabwc, for diabetes with complications;

hp, for hemiplegia or paraplegia;

rend, for renal disease;

canc, for cancer (any malignancy);

msld, for moderate or severe liver disease;

metacanc, for metastatic solid tumour;

aids, for AIDS/HIV;

Conversely, for the Elixhauser score the dataset contains the following variables:

The id variable as defined by the user;

chf, for congestive heart failure;

carit, for cardiac arrhythmias;

valv, for valvular disease;

pcd, for pulmonary circulation disorders;

pvd, for peripheral vascular disorders;

hypunc, for hypertension, uncomplicated;

hypc, for hypertension, complicated;

para, for paralysis;

ond, for other neurological disorders;

cpd, for chronic pulmonary disease;

diabunc, for diabetes, uncomplicated;

diabc, for diabetes, complicated;

hypothy, for hypothyroidism;

rf, for renal failure;

ld, for liver disease;

pud, for peptic ulcer disease, excluding bleeding;

aids, for AIDS/HIV;

lymph, for lymphoma;

metacanc, for metastatic cancer;

solidtum, for solid tumour, without metastasis;

rheumd, for rheumatoid arthritis/collaged vascular disease;

coag, for coagulopathy;

obes, for obesity;

wloss, for weight loss;

fed, for fluid and electrolyte disorders;

blane, for blood loss anaemia;

dane, for deficiency anaemia;

alcohol, for alcohol abuse;

drug, for drug abuse;

psycho, for psychoses;

depre, for depression;

Labels are presented to the user when using the RStudio viewer (e.g. via the utils::View() function) for convenience.

Details

The ICD-10 and ICD-9-CM coding for the Charlson and Elixhauser scores is based on work by Quan et al. (2005). ICD-10 and ICD-9 codes must be in upper case and with alphanumeric characters only in order to be properly recognised; set tidy.codes = TRUE to properly tidy the codes automatically. A message is printed to the R console when non-alphanumeric characters are found.

References

Quan H, Sundararajan V, Halfon P, Fong A, Burnand B, Luthi JC, et al. Coding algorithms for defining comorbidities in ICD-9-CM and ICD-10 administrative data. Medical Care 2005; 43(11):1130-1139.

Charlson ME, Pompei P, Ales KL, et al. A new method of classifying prognostic comorbidity in longitudinal studies: development and validation. Journal of Chronic Diseases 1987; 40:373-383.

Ludvigsson JF, Appelros P, Askling J et al. Adaptation of the Charlson Comorbidity Index for register-based research in Sweden. Clinical Epidemiology 2021; 13:21-41.

Sundararajan V, Henderson T, Perry C, Muggivan A, Quan H, Ghali WA. New ICD-10 version of the Charlson comorbidity index predicted in-hospital mortality. Journal of Clinical Epidemiology 2004; 57(12):1288-1294.

Examples

set.seed(1)
x <- data.frame(
  id = sample(1:15, size = 200, replace = TRUE),
  code = sample_diag(200),
  stringsAsFactors = FALSE
)

# Charlson score based on ICD-10 diagnostic codes:
comorbidity(x = x, id = "id", code = "code", map = "charlson_icd10_quan", assign0 = FALSE)

#>    id ami chf pvd cevd dementia copd rheumd pud mld diab diabwc hp rend canc
#> 1   1   0   0   0    0        0    0      0   0   0    0      0  0    0    1
#> 2   2   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 3   3   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 4   4   0   0   0    0        0    0      0   0   0    0      0  0    0    1
#> 5   5   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 6   6   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 7   7   0   1   0    0        0    0      0   0   1    0      0  0    0    1
#> 8   8   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 9   9   0   0   0    1        0    0      0   0   0    0      0  0    0    1
#> 10 10   0   0   0    0        0    0      0   0   0    0      0  0    0    1
#> 11 11   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 12 12   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 13 13   0   0   0    0        0    0      0   1   0    0      0  0    0    1
#> 14 14   0   0   0    0        0    0      0   0   0    0      0  0    0    0
#> 15 15   0   0   0    0        0    0      0   0   0    0      0  0    0    1
#>    msld metacanc aids
#> 1     0        0    0
#> 2     0        0    1
#> 3     0        0    0
#> 4     0        0    0
#> 5     0        0    0
#> 6     0        0    0
#> 7     0        0    0
#> 8     0        0    0
#> 9     0        0    0
#> 10    0        0    0
#> 11    0        0    0
#> 12    0        0    0
#> 13    0        0    0
#> 14    0        0    0
#> 15    0        0    0

# Elixhauser score based on ICD-10 diagnostic codes:
comorbidity(x = x, id = "id", code = "code", map = "elixhauser_icd10_quan", assign0 = FALSE)

#>    id chf carit valv pcd pvd hypunc hypc para ond cpd diabunc diabc hypothy rf
#> 1   1   0     0    0   0   0      0    0    0   0   0       0     0       1  0
#> 2   2   0     0    1   0   0      0    0    0   0   0       0     0       0  0
#> 3   3   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 4   4   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 5   5   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 6   6   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 7   7   1     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 8   8   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 9   9   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 10 10   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 11 11   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 12 12   0     0    0   0   0      0    0    0   1   0       0     0       0  0
#> 13 13   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 14 14   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#> 15 15   0     0    0   0   0      0    0    0   0   0       0     0       0  0
#>    ld pud aids lymph metacanc solidtum rheumd coag obes wloss fed blane dane
#> 1   0   0    0     0        0        1      0    0    0     0   0     0    0
#> 2   0   0    1     0        0        0      0    0    0     0   0     0    0
#> 3   0   0    0     0        0        0      0    0    0     0   0     0    0
#> 4   0   0    0     0        0        0      0    0    0     0   0     0    0
#> 5   0   0    0     0        0        0      0    0    0     0   0     0    0
#> 6   0   0    0     0        0        0      0    0    0     0   0     0    0
#> 7   1   0    0     0        0        1      0    0    0     0   0     0    0
#> 8   0   0    0     0        0        0      0    0    0     0   0     0    0
#> 9   0   0    0     0        0        1      1    0    0     0   0     0    0
#> 10  0   0    0     0        0        1      0    0    0     0   0     0    0
#> 11  0   0    0     0        0        0      0    0    0     0   0     0    0
#> 12  0   0    0     0        0        0      0    1    0     0   0     0    0
#> 13  0   1    0     0        0        1      0    0    0     0   0     0    0
#> 14  0   0    0     0        0        0      0    0    0     0   0     0    0
#> 15  0   0    0     0        0        1      0    0    0     0   0     0    0
#>    alcohol drug psycho depre
#> 1        0    0      0     0
#> 2        0    0      0     0
#> 3        0    0      0     0
#> 4        0    0      0     0
#> 5        0    0      0     0
#> 6        0    0      0     0
#> 7        1    0      0     0
#> 8        0    0      0     0
#> 9        0    0      0     0
#> 10       0    0      0     0
#> 11       0    0      0     0
#> 12       0    1      1     0
#> 13       0    0      0     0
#> 14       0    0      0     0
#> 15       0    0      0     0