comorbidity: Computing Comorbidity Scores

Computing comorbidity indices and scores such as the weighted Charlson score (Charlson, 1987 doi:10.1016/0021-9681(87)90171-8 ) and the Elixhauser comorbidity score (Elixhauser, 1998 doi:10.1097/00005650-199801000-00004 ) using ICD-9-CM or ICD-10 codes (Quan, 2005 doi:10.1097/01.mlr.0000182534.19832.83 ). Australian and Swedish modifications of the Charlson Comorbidity Index are available as well (Sundararajan, 2004 doi:10.1016/j.jclinepi.2004.03.012 and Ludvigsson, 2021 doi:10.2147/CLEP.S282475 ), together with different weighting algorithms for both the Charlson and Elixhauser comorbidity scores.

See also

Author

Maintainer: Alessandro Gasparini alessandro@ellessenne.xyz (ORCID)

Other contributors:

Hojjat Salmasian hsalmasian@bwh.harvard.edu (ICD-9-CM scores) [contributor]

Jonathan Williman jonathan.williman@otago.ac.nz (ORCID) [contributor]

Sing Yi Chia cxycsy@hotmail.com (ORCID) [contributor]

Edmund Teo ed_teo8@yahoo.com.sg (ORCID) [contributor]

Desi Quintans [contributor]